requirements.txt
python-igraph==0.9.8
thefuzz==0.19.0